Implement submission importer
Allows the user to press on a new "import directory" button on the file
submissions page and choose a directory to import from. The directory
will be inspected and an appropriate importer chosen automatically. The
submissions will then be imported in logical order, potentially
including thumbnails, titles, descriptions etc.

Currently the following importers are implemented, because they are the
ones I need:

* A flat file importer. Just gobbles up all the files in a directory.

* An importer for ActivityPub exports, like they fall out of Mastodon
  for example. Grabs all media posts that aren't direct messages and
  turns them into submissions with descriptions and tags.

* An importer for [this FurAffinity
  archiver](https://github.com/askmeaboutlo0m/faescape), which will
  import submissions with titles, descriptions, tags, ratings and
  thumbnails.

Further importers can be added pretty easily by extending `Importer`
and adding an entry in the `SubmissionImporterService.importers` array.
The importers `inspect` the directory and decide if they're equipped to
handle them. The first matching importer will be chosen to do the
extraction.

The progress reporting is kinda ugly with periodic messages in the
corner, but it does the job.

Implements #170.