Sensibly check keep defaults when loading template
When editing multiple submissions, the state of the keep defaults
checkbox is now guessed from what's being loaded and what's been input.
If a default part is explicitly being loaded from the template or
there's alredy something filled into the form, it's checked off,
otherwise on. This avoids a user loading a websites-only template and
then obliterating the default part on all the submissions they're
editing (like I did several times.) This is again just a sensible
default and can be explicitly toggled if the user desires.