Add parent submission field for reply threads
This adds a new "parent submission" field to the default submission
section, letting the user pick a submission's parent from a drop-down.
After posting a submission, website services get a chance to react to
this and manipulate the child's part however they feel is useful.

Currently, this is implemented for Bluesky, Mastodon and Pleroma, which
will set the "reply to post URL" field to the source URL of the parent
post, removing the necessity to manually copy those over.

Other uses could be implemented if someone feels like it, different
websites have various flavors of post relationships that could be
handled this way.

The only time the application acts on parent-child relationships is
right after a submission has been posted. That means invalid parents
that reference removed submissions cause no harm, they're presented to
the user as being unassigned. That means we don't need to bother
maintaining the validity of these fields when removing submissions. We
also don't need to check for circular dependencies, since those circles
will break themselves as soon as one submission out of them is posted.