Allow automatic insertion of parent source links
When a submission is posted, it will now look through all child
descriptions and replace instances of `{parent:WHAT}` with a source URL.
`WHAT` can either be an internal website name or, if that would be
ambiguous, an account ID. The shortcut description shows the user a list
of options they can copy from. The account list page also shows these
account ids now.